Rewrite eval "loop" with gotos
I was having difficulty expressing the reduction rules in a loop; the
tricky one is that when a function has been applied, the next thing to
do is always to look at the continuation (stack frame). Otherwise,
expressions that result in lists might get evaluated.

The easiest way to get this right is to just always jump to the next
thing to do.