Put forward pointers in the header
This is useful if the actual storage of an obj takes up no room, e.g.,
a zero-length string, since that means there's no place to put the
pointer *after* the header.

The trick is to be able to distinguish between a regular header and a
forward pointer. To do this, we use the lower bits of the header for
the tag, and make sure the regular tags set at least one of the lower
bits, which will be zero for pointers.