Short symbols become intermediate values
If a symbol's characters (and null terminator) will fit in the bytes
of a pointer-sized value, we just put them there and give it an
immediate tag.

Because I'm punning a number/pointer as a char*, I need to be careful
where I put the characters in the number since I'll be masking off
some of it, and the tag might occur at the beginning or end depending
on the endianness. I've made it work for little-endian; I might need