Make primitives immediate and give them names
Since these are pointers, I can set one of the lower bits to
distinguish them as immediate values.

To give primitives names I put them in a struct, which means an
indirect to apply; otherwise, I might just have made the function
pointer the immediate value.

I also changed the way syms are tagged: since they don't care about
the last byte of the value (it can't be used as a character because of
the tag), they can use more of the lower bits as tag.