Use arg continuation to implement letrec
letrec needs three pieces of information: the environment in which
assignments (and the body) will be evaluated, the assignments
themselves, and body. But my frames really have only one place to put
things, `f->args`.

Therefore, I use another continuation frame to calculate each arg and
put it in the correct place in the environment (a vector), then put
that vector back in the val register. The same arg frames can be used
for calculating the arguments to a function (hence the name).