Use a stack of obj rather than frame structs
Each kind of continuation tends to need slightly different things put
on the stack. Instead of having a frame struct that is used for all
these purposes, just put values on the stack, including numbers
representing the continuations.

A downside is that it's hard to tell the structure of the stack by
looking at the values, since the frame boundaries look like numbers.

Limitations: does not currently store/restore the environment; tends
to unwind continuations then wind them back onto the stack (rather
than peeking/poking at the contents and possibly unwinding).