Save and restore the environment when entering a new context
When the interpreter enters a closure or letrec, it needs to swap the
current environment for another. If it's not a tail-call, the
environment at that time will need to be restored.

I've implemented the restoration by pushing a "return" frame onto the
stack. The easiest place to do this is when I'm examining the next
expression to see if it's a special form. I know whether it's a
tailcall or not from the evaluation (only expressions in a "body", or
arguments, can be non-tailcall).